Document improved (#9)
This pull request updates the `README.md` file to provide clearer
instructions for running the Matrix Rotator application. The main change
is the addition of instructions for running pre-built executable
programs and updating the command for running the application directly.

Changes to `README.md`:

* Added instructions for launching pre-built executable programs,
including the command format.
* Updated the command for running the application directly to include
input and output path parameters.